requirements.txt
ultralytics
flask
waitress
pillow
gunicorn
opencv-python
scikit-image
numpy